% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_formal_context.R
\name{PartialFormalContext}
\alias{PartialFormalContext}
\title{R6 class for a partial formal context}
\description{
This class implements the data structure and methods for partial formal contexts.
}
\section{Public fields}{

\itemize{
\item \code{I}: the table for the partial formal contexts.
\item \code{attributes}: name of the attributes in the partial formal contexts.
\item \code{objects}: name of the objects in the context.
\item \code{grades_set}: set of grades ( \{0, -1, 1\}) of the attributes.
\item \code{concepts}: list of concepts (extent, intent).
\item \code{implications}: extracted implications as an \code{ImplicationSet}.
}
}

\examples{
# Build and print the formal context
fc_planets <- PartialFormalContext$new(planets)
print(fc_planets)

# Define a set of attributes
S <- Set$new(attributes = fc_planets$attributes)
S$assign(moon = 1, large = 1)

# Compute the closure of S
Sc <- fc_planets$closure(S)
# Is Sc a closed set?
fc_planets$is_closed(Sc)

# Clarify and reduce the formal context
fc2 <- fc_planets$reduce(TRUE)

# Find implications
fc_planets$find_implications()

# Read a formal context from CSV
filename <- system.file("contexts", "airlines.csv", package = "fcaR")
fc <- PartialFormalContext$new(filename)

# Read a formal context from a CXT file
filename <- system.file("contexts", "lives_in_water.cxt", package = "fcaR")
fc <- PartialFormalContext$new(filename)


## ------------------------------------------------
## Method `PartialFormalContext$incidence`
## ------------------------------------------------

fc <- PartialFormalContext$new(planets)
fc$incidence()
}
\references{
Guigues J, Duquenne V (1986). “Familles minimales d'implications informatives résultant d'un tableau de données binaires.” \emph{Mathématiques et Sciences humaines}, \emph{95}, 5-18.

Ganter B, Wille R (1999). \emph{Formal concept analysis : mathematical foundations}. Springer. ISBN 3540627715.

Belohlavek R (2002). “Algorithms for fuzzy concept lattices.” In \emph{Proc. Fourth Int. Conf. on Recent Advances in Soft Computing}. Nottingham, United Kingdom, 200-205.

Hahsler M, Grun B, Hornik K (2005). “arules - a computational environment for mining association rules and frequent item sets.” \emph{J Stat Softw}, \emph{14}, 1-25.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{I}}{The table of the formal context as a matrix.}

\item{\code{attributes}}{The attributes of the formal context.}

\item{\code{objects}}{The objects of the formal context}

\item{\code{grades_set}}{The set of degrees (in [0,-1, 1]) the whole set of attributes can take.}

\item{\code{concepts}}{The concept lattice associated to the formal context as a \code{\link{ConceptLattice}}.}

\item{\code{implications}}{A set of implications on the formal context as an \code{\link{ImplicationSet}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PartialFormalContext-new}{\code{PartialFormalContext$new()}}
\item \href{#method-PartialFormalContext-is_empty}{\code{PartialFormalContext$is_empty()}}
\item \href{#method-PartialFormalContext-intent}{\code{PartialFormalContext$intent()}}
\item \href{#method-PartialFormalContext-extent}{\code{PartialFormalContext$extent()}}
\item \href{#method-PartialFormalContext-closure}{\code{PartialFormalContext$closure()}}
\item \href{#method-PartialFormalContext-obj_concept}{\code{PartialFormalContext$obj_concept()}}
\item \href{#method-PartialFormalContext-att_concept}{\code{PartialFormalContext$att_concept()}}
\item \href{#method-PartialFormalContext-is_concept}{\code{PartialFormalContext$is_concept()}}
\item \href{#method-PartialFormalContext-is_closed}{\code{PartialFormalContext$is_closed()}}
\item \href{#method-PartialFormalContext-find_concepts}{\code{PartialFormalContext$find_concepts()}}
\item \href{#method-PartialFormalContext-find_implications}{\code{PartialFormalContext$find_implications()}}
\item \href{#method-PartialFormalContext-process_implications}{\code{PartialFormalContext$process_implications()}}
\item \href{#method-PartialFormalContext-load}{\code{PartialFormalContext$load()}}
\item \href{#method-PartialFormalContext-dim}{\code{PartialFormalContext$dim()}}
\item \href{#method-PartialFormalContext-print}{\code{PartialFormalContext$print()}}
\item \href{#method-PartialFormalContext-to_latex}{\code{PartialFormalContext$to_latex()}}
\item \href{#method-PartialFormalContext-incidence}{\code{PartialFormalContext$incidence()}}
\item \href{#method-PartialFormalContext-plot}{\code{PartialFormalContext$plot()}}
\item \href{#method-PartialFormalContext-clone}{\code{PartialFormalContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-new}{}}}
\subsection{Method \code{new()}}{
Creator for the Partial Formal Context class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$new(I, filename, remove_const = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{I}}{(numeric matrix) The table of the partial formal context.}

\item{\code{filename}}{(character) Path of a file to import.}

\item{\code{remove_const}}{(logical) If \code{TRUE}, remove constant columns. The default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Columns of \code{I} should be named, since they are the names of the attributes of the partial formal context.

If no \code{I} is used, the resulting \code{PartialFormalContext} will be empty and not usable unless for loading a previously saved one. In this case, one can provide a \code{filename} to import. Only RDS, CSV and CXT files are currently supported.
}

\subsection{Returns}{
An object of the \code{PartialFormalContext} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-is_empty"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-is_empty}{}}}
\subsection{Method \code{is_empty()}}{
Check if the \code{PartialFormalContext} is empty
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$is_empty()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the \code{PartialFormalContext} is empty, that is, has not been provided with a matrix, and \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-intent"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-intent}{}}}
\subsection{Method \code{intent()}}{
Get the intent of a partial set of objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$intent(S)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{S}}{(\code{Set}) The set of objects to compute the intent for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Set} with the intent.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-extent"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-extent}{}}}
\subsection{Method \code{extent()}}{
Get the extent of a partial set of attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$extent(S)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{S}}{(\code{Set}) The set of attributes to compute the extent for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Set} with the intent.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-closure"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-closure}{}}}
\subsection{Method \code{closure()}}{
Get the closure of a partial set of attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$closure(S)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{S}}{(\code{Set}) The set of attributes to compute the closure for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Set} with the closure.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-obj_concept"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-obj_concept}{}}}
\subsection{Method \code{obj_concept()}}{
Object Concept
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$obj_concept(object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{(character) Name of the object to compute its associated concept}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object concept associated to the object given.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-att_concept"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-att_concept}{}}}
\subsection{Method \code{att_concept()}}{
Attribute Concept
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$att_concept(attribute, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attribute}}{(character) Name of the attribute to compute its associated concept}

\item{\code{value}}{(numerical) Value for the attribute, It can be 1 or -1.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The attribute concept associated to the attribute given.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-is_concept"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-is_concept}{}}}
\subsection{Method \code{is_concept()}}{
Is a Concept?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$is_concept(C)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{C}}{A \code{Concept} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if \code{C} is a concept.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-is_closed"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-is_closed}{}}}
\subsection{Method \code{is_closed()}}{
Testing closure of attribute sets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$is_closed(S)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{S}}{A \code{Set} of attributes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if the set \code{S} is closed in this formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-find_concepts"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-find_concepts}{}}}
\subsection{Method \code{find_concepts()}}{
Use Ganter Algorithm to compute concepts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$find_concepts(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(logical) TRUE will provide a verbose output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with all the concepts in the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-find_implications"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-find_implications}{}}}
\subsection{Method \code{find_implications()}}{
Use Ganter Algorithm to compute implications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$find_implications(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(logical) TRUE will provide a verbose output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with all the implications in the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-process_implications"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-process_implications}{}}}
\subsection{Method \code{process_implications()}}{
Get the extent of a partial set of attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$process_implications(S)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{S}}{(\code{Set}) The set of attributes to compute the extent for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Set} with the intent.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-load"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-load}{}}}
\subsection{Method \code{load()}}{
Load a \code{PartialFormalContext} from a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$load(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{(character) Path of the file to load the \code{PartialFormalContext} from.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Currently, only RDS, CSV and CXT files are supported.
}

\subsection{Returns}{
The loaded \code{PartialFormalContext}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-dim"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-dim}{}}}
\subsection{Method \code{dim()}}{
Dimensions of the formal context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with (number of objects, number of attributes).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-print"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-print}{}}}
\subsection{Method \code{print()}}{
Prints the formal context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prints information regarding the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-to_latex"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-to_latex}{}}}
\subsection{Method \code{to_latex()}}{
Write the context in LaTeX format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$to_latex(
  table = TRUE,
  label = "",
  caption = "",
  fraction = c("none", "frac", "dfrac", "sfrac")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{(logical) If \code{TRUE}, surrounds everything between \code{\\begin{table}} and \code{\\end{table}}.}

\item{\code{label}}{(character) The label for the table environment.}

\item{\code{caption}}{(character) The caption of the table.}

\item{\code{fraction}}{(character) If \code{none}, no fractions are produced. Otherwise, if it is \code{frac}, \code{dfrac} or \code{sfrac}, decimal numbers are represented as fractions with the corresponding LaTeX typesetting.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A table environment in LaTeX.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-incidence"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-incidence}{}}}
\subsection{Method \code{incidence()}}{
Incidence matrix of the partial formal context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$incidence()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The incidence matrix of the partial formal context
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fc <- PartialFormalContext$new(planets)
fc$incidence()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the formal context table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$plot(to_latex = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{to_latex}}{(logical) If \code{TRUE}, export the plot as a \code{tikzpicture} environment that can be included in a \code{LaTeX} file.}

\item{\code{...}}{Other parameters to be passed to the \code{tikzDevice} that renders the lattice in \code{LaTeX}, or for the figure caption. See \code{Details}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Particular parameters that control the size of the \code{tikz} output are: \code{width}, \code{height} (both in inches), and \code{pointsize} (in points), that should be set to the font size used in the \code{documentclass} header in the \code{LaTeX} file where the code is to be inserted.

If a \code{caption} is provided, the whole \code{tikz} picture will be wrapped by a \code{figure} environment and the caption set.
}

\subsection{Returns}{
If \code{to_latex} is \code{FALSE}, it returns nothing, just plots the graph of the formal context. Otherwise, this function returns the \code{LaTeX} code to reproduce the formal context plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartialFormalContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PartialFormalContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartialFormalContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
